rename directory to v1
